More Bug Progress
Made some more progress regarding the bug. A few logic operators were out of place. Re organized. Heavy arrow mesh is spawn but not attaching to bow, while light arrow actor is invisible, but drawing the bow. The display number is only correct when the trigger is pressed down as well. Steady progress.