Created Functions & Updated Deprecated Nodes
Created functions to check for current health status, created function to check for arrow type and validity. Detirmined the arrow isn't actually spawning for whatever reason. Breakpoints show that the arrow is spawning, and that the heavy arrow variable, and the light arrow variable are being set to instances of the objects. Updated deprecated attach to component nodes, including location, rotation, and scale enum arguments.  Arrow's still no appearing when spawned, and arguments state that variables have an object when arrow shoot function is being called and during detach node.