Finished Tutorial & Fixed Endgame Audio bug
Managed to use the end wave array count instead of wave struct array count to act as the No of wave condition to the game over function. The game over function still needs to be placed in a few places (Check for PLayer health == 0, and Castle Health == 0) Tutorial is 100% complete and playable.