Grad, Beam, & Tutorial Changes
Recieved feedback regarding UI/UX being difficult for the player to learn how to play, & that iteration is required. Bow feels good. Updated the beam function to get rid of a null error when referring to the heavy arrows. Updated the tutorial level to be shorter. Discovered a bug with the bow that spawns the beam before the arrow is attached to the bow when the arrow comes in extremely close proximity to the bow is pulled away and the trigger is pulled. The beam still functions as it would if the arrow was attached to the bow.